Move filesystem helpers to source file and detail include
