+++ libs/pika/runtime_configuration/include/pika/runtime_configuration/init_ini_data.hpp
@@ -7,8 +7,8 @@
 
 #pragma once
 
+#include <pika/detail/filesystem.hpp>
 #include <pika/ini/ini.hpp>
-#include <pika/modules/filesystem.hpp>
 
 #include <map>
 #include <memory>
